chore: move reusable utils into sam_utils
